requirements.txt
APScheduler==3.10.4
certifi==2023.11.17
charset-normalizer==3.3.2
greenlet==3.0.1
idna==3.6
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
stripe==7.6.0
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
